Add files via upload
This zip file contains all the datasets used for the research carried out in this thesis. This consists of: the initial dataset (Clark 2024), the corpus, the sampled data, the annotated data. The annotated dataset contains thematic annotations for songs from Taylor Swift's discography. These datasets can be used for research purposes.